+++ test/unit/fuzz-tools.test.js
@@ -488,6 +488,7 @@ import * as banffTcmrV510 from '../../lib/banff-tcmr-v510.js';
 import * as crafftV511 from '../../lib/crafft-v511.js';
 import * as vaizeyV512 from '../../lib/vaizey-v512.js';
 import * as asrsV513 from '../../lib/asrs-v513.js';
+import * as ymrsV514 from '../../lib/ymrs-v514.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -921,6 +922,7 @@ const MODULES = {
   'crafft-v511.js': crafftV511,
   'vaizey-v512.js': vaizeyV512,
   'asrs-v513.js': asrsV513,
+  'ymrs-v514.js': ymrsV514,
   'tb-testing.js': tbTesting,
 };